Fix compilation error on windows+msvc2013
- Some missing includes of <algorithms> in order to have
  stdd::min() and similar functions.
- Need to cast numIndices and numNVerts to int explicitly
  in order to solve warning treated as an error about
  precision loss.
- Can't do vector[0] for an empty vector, it'll generate
  a runtime range check error.
- MSVC only works fine with make_pair(foo, bar) syntax,
  without explicit template substitution here. Otherwise
  weird 'can't cast int to int&&' errors are happening.

+++ opensubdiv/osdutil/evaluator_capi.cpp
@@ -161,9 +161,9 @@ void openSubdiv_getEvaluatorTopology(
     // application to test this yet.
     *numVertices =  evaluation_descr->topology.numVertices;
     *refinementLevel =  evaluation_descr->topology.refinementLevel;
-    *numIndices =  evaluation_descr->topology.indices.size();
+    *numIndices =  (int)evaluation_descr->topology.indices.size();
     *indices =  &evaluation_descr->topology.indices[0];
-    *numNVerts =  evaluation_descr->topology.nverts.size();
+    *numNVerts =  (int)evaluation_descr->topology.nverts.size();
     *nverts =  &evaluation_descr->topology.nverts[0];
 }
